Fixed non unity compiler issues and reverted a couple of changes
