+++ Gems/Atom/Tools/AtomToolsFramework/Code/Source/Util/MaterialPropertyUtil.cpp
@@ -14,6 +14,10 @@
 #include <AtomToolsFramework/DynamicProperty/DynamicProperty.h>
 #include <AtomToolsFramework/Util/MaterialPropertyUtil.h>
 #include <AtomToolsFramework/Util/Util.h>
+#include <AzCore/Math/Color.h>
+#include <AzCore/Math/Vector2.h>
+#include <AzCore/Math/Vector3.h>
+#include <AzCore/Math/Vector4.h>
 
 namespace AtomToolsFramework
 {
@@ -123,6 +127,41 @@ namespace AtomToolsFramework
         {
             propertyMetaData.m_visibility = AZ::RPI::MaterialPropertyVisibility::Enabled;
         }
+    }
+
+    template<typename T>
+    bool ComparePropertyValues(const AZStd::any& valueA, const AZStd::any& valueB)
+    {
+        return valueA.is<T>() && valueB.is<T>() && *AZStd::any_cast<T>(&valueA) == *AZStd::any_cast<T>(&valueB);
+    }
+
+    bool ArePropertyValuesEqual(const AZStd::any& valueA, const AZStd::any& valueB)
+    {
+        if (valueA.type() != valueB.type())
+        {
+            return false;
+        }
+
+        if (ComparePropertyValues<bool>(valueA, valueB) ||
+            ComparePropertyValues<int32_t>(valueA, valueB) ||
+            ComparePropertyValues<uint32_t>(valueA, valueB) ||
+            ComparePropertyValues<float>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Vector2>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Vector3>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Vector4>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Color>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Data::AssetId>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Data::Asset<AZ::Data::AssetData>>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Data::Asset<AZ::RPI::ImageAsset>>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Data::Asset<AZ::RPI::StreamingImageAsset>>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Data::Asset<AZ::RPI::MaterialAsset>>(valueA, valueB) ||
+            ComparePropertyValues<AZ::Data::Asset<AZ::RPI::MaterialTypeAsset>>(valueA, valueB) ||
+            ComparePropertyValues<AZStd::string>(valueA, valueB))
+        {
+            return true;
+        }
+
+        return false;
     }
 
     bool ConvertToExportFormat(
